1차 pr
