+++ package.json
@@ -18,7 +18,6 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.7.7",
-    "axios": "^1.7.7",
     "js-cookie": "^3.0.5",
     "jwt-decode": "^4.0.0",
     "mui-tel-input": "^6.0.1",
@@ -56,6 +55,7 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@eslint/js": "^9.10.0",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
