+++ src/context/SidebarContext.js
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useState, useEffect } from "react";
+import React, { createContext, useContext, useState } from "react";
 
 const SidebarContext = createContext();
