+++ src/pages/chat/Allchat.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Allchat() {
+  return <div>전체 채팅 페이지</div>;
+}
