+++ src/shared/ProtectedRoute.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Route, Navigate } from "react-router-dom";
+import { Navigate } from "react-router-dom";
 
 const ProtectedRoute = ({ element, isLoggedIn }) => {
   return isLoggedIn ? element : <Navigate to="/" />;
