+++ src/shared/Router.js
@@ -1,8 +1,10 @@
 import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { Routes, Route, Navigate } from "react-router-dom";
 
 // 홈 페이지
 import Home from "../pages/Home";
+
+import Login from "../pages/login/Login";
 
 // 게시판 페이지
 import FreeBoard from "../pages/board/FreeBoard";
@@ -24,32 +26,49 @@ import Vote from "../pages/lectureRoom/Vote";
 
 // 채팅 페이지
 import Chat from "../pages/chat/Chat";
+import Allchat from "../pages/chat/Allchat";
 import Privatechat from "../pages/chat/Privatechat";
 
 // 캘린더 페이지
 import Calendar from "../pages/calendar/Calendars";
 
-const Router = () => {
+const Router = ({ isLoggedIn }) => {
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route element={<LayoutWrapper />}>
-          <Route path="/" element={<Home />} />
-          <Route path="freeboard" element={<FreeBoard />} />
-          <Route path="noticeboard" element={<NoticeBoard />} />
-          <Route path="board" element={<Board />} />
-          <Route path="profile" element={<Profile />} />
-          <Route path="PasswordCheck" element={<PasswordCheck />} />
-          <Route path="lectureroom" element={<LectureRoom />} />
-          <Route path="chat" element={<Chat />} />
-          <Route path="calendar" element={<Calendar />} />
-          <Route path="assignment" element={<Assignment />} />
-          <Route path="qna" element={<Qna />} />
-          <Route path="vote" element={<Vote />} />
-          <Route path="privatechat" element={<Privatechat />} />
-        </Route>
-      </Routes>
-    </BrowserRouter>
+    <Routes>
+      <Route element={<LayoutWrapper />}>
+        {/* 홈 */}
+        <Route path="/" element={<Home />} />
+
+        {/* 로그인 폼 */}
+        <Route
+          path="login"
+          element={isLoggedIn ? <Navigate to="/" /> : <Login />} // 로그인 상태일 경우 홈으로 리다이렉트
+        />
+
+        {/* 게시판 */}
+        <Route path="board" element={<Board />} />
+        <Route path="freeboard" element={<FreeBoard />} />
+        <Route path="noticeboard" element={<NoticeBoard />} />
+
+        {/* 유저 관련 페이지 */}
+        <Route path="profile" element={<Profile />} />
+        <Route path="PasswordCheck" element={<PasswordCheck />} />
+
+        {/* 강의실 페이지 */}
+        <Route path="lectureroom" element={<LectureRoom />} />
+        <Route path="assignment" element={<Assignment />} />
+        <Route path="qna" element={<Qna />} />
+        <Route path="vote" element={<Vote />} />
+
+        {/* 채팅 페이지 */}
+        <Route path="chat" element={<Chat />} />
+        <Route path="allchat" element={<Allchat />} />
+        <Route path="privatechat" element={<Privatechat />} />
+
+        {/* 캘린더 */}
+        <Route path="calendar" element={<Calendar />} />
+      </Route>
+    </Routes>
   );
 };
